refactor: split state.py into user_data.py and settings.py
Move GUIColors and GUISettings to settings.py. Move GUIInput and GUIState to user_data.py. Update all GUI view and unit test imports to match.
