+++ src/amplifyp/gui/views/input_view.py
@@ -12,7 +12,8 @@ from typing import Any
 
 import flet as ft
 
-from amplifyp.gui.state import GUIColors, GUIState
+from amplifyp.gui.settings import GUIColors
+from amplifyp.gui.user_data import GUIState
 from amplifyp.gui.util import clean_sequence, format_sequence
